% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam.init.r
\name{gam.init}
\alias{gam.init}
\title{Function to select initial gamma values spanning its space}
\usage{
    gam.init(n.initials, Z, lb.quantile, ub.quantile, ss=1)
}
\arguments{
    \item{n.initials}{Number of gamma values.}
    \item{Z}{The grouping variables.}
    \item{lb.quantile}{The lower quantile specified for Z\%*\%gamma.initials.}
    \item{ub.quantile}{The upper quantile specified for Z\%*\%gamma.initials.}
    \item{ss}{A positive integer with n.initials/ss indicating how many sets of gamma are chosen, default to 1.}
}
\value{
    A matrix. A set of gamma values selected for defining subgroup.
}
\examples{
    n = 200
    m = 4
    p1 = 2
    p2 = 1
    p3 = 3
    alpha = rep(1, p1)
    beta  = rep(1, p2)/2
    gamma = c(1, seq(-1,1,length.out = p3-1)) 
    rho = 0
    set.seed(100)
    data = generate_data(n, m, alpha, beta, gamma, rho)
    K = 1000
    qlb = 0.1
    cols = apply(data$U, 2, var) != 0
    Gamma = gam.init(K, data$U[,cols], lb.quantile=qlb, ub.quantile=1-qlb, ss=1)
}
