% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.r
\name{generate_data}
\alias{generate_data}
\title{Function for generating longitudinal data from a change plane model.}
\usage{
    generate_data(n, m, alpha, beta, gamma, rho)
}
\arguments{
    \item{n}{A constant. The sample size.}
    \item{m}{A constant. The number of time point.}
    \item{alpha}{A vector. The true parameter for baseline covariates.}
    \item{beta}{A vector. The true parameter denoting the heterogeneous effect of the subgroup.}
    \item{gamma}{A vector. The true parameter for grouping variables.}
    \item{rho}{The strength of correlation among covariates.}
}
\value{
A list.
\itemize{
    \item Y - A \eqn{n*m} matrix. The response.
    \item X - A \eqn{(n*m)*p} matrix. The baseline covariates.
    \item Z - A \eqn{(n*m)*q} matrix. The baseline covariates.
    \item U - A \eqn{(n*m)*r} matrix. The grouping variables.
}
}
\examples{
    n = 200
    m = 4
    p1 = 2
    p2 = 1
    p3 = 3
    alpha = rep(1, p1)
    beta  = rep(1, p2)/2
    gamma = c(1, seq(-1,1,length.out = p3-1)) 
    rho = 0
    set.seed(100)
    data = generate_data(n, m, alpha, beta, gamma, rho)
}
