% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongSubTestWAST.r
\name{LongSubTestWAST}
\alias{LongSubTestWAST}
\title{The weighted average of the score test for the change-plane model with longitudinal data.}
\usage{
    LongSubTestWAST(data, family = 1, B = 1000, seed = 1)
}
\arguments{
    \item{data}{A list, including \eqn{Y} (n\%*\%m response), \eqn{X} ((n\%*\%m)\%*\%p predictors), \eqn{Z} ((n\%*\%m)\%*\%q predictors), \eqn{U} ((n\%*\%m)\%*\%r grouping variables).}
    \item{family}{The corresponding family. family = 1 denotes the Gaussian family, and family = 2 denotes the binomial family. Default is 1.}
    \item{B}{A constant. Number of bootstrap samples. Default is 1000.}
    \item{seed}{A constant. The number of seeds for generating the initial gamma values. Default is 1.}
}
\value{
A list.
\itemize{
    \item tn0 - The value of test statistic.
    \item pval - The p-value of the test.
    \item time - Running time.
}
}
\references{Novel subgroup testing for longitudinal data.}
\examples{
    n = 200
    m = 4
    p1 = 2
    p2 = 1
    p3 = 3
    alpha = rep(1, p1)
    beta  = rep(1, p2)/2
    gamma = c(1, seq(-1,1,length.out = p3-1)) 
    rho = 0
    set.seed(100)
    data = generate_data(n, m, alpha, beta, gamma, rho)
    fit <- LongSubTestWAST(data)
}